requirements.txt
anthropic
python-pptx
